+++ src/components/Bookmarks.vue
@@ -5,7 +5,7 @@
       <a :href="getlink(bookmark)" @click.prevent="showsensor(bookmark)">
         <b v-if="bookmark?.customName" class="name">{{ bookmark.customName }}</b>
         <b v-if="bookmark?.geo" :class="bookmark.customName ? 'addresssm' : 'adresslg'">
-          {{JSON.parse(bookmark.geo)?.lat}}, {{JSON.parse(bookmark.geo)?.lng}}
+          {{ JSON.parse(bookmark.geo)?.lat }}, {{ JSON.parse(bookmark.geo)?.lng }}
         </b>
       </a>
       <button title="Remove this sensor" @click.prevent="deletebookmark(bookmark.id)">
@@ -15,75 +15,65 @@
   </div>
 </template>
 
-<script>
-import { useStore } from "@/store";
+<script setup>
+import { computed, onMounted } from "vue";
+import { useBookmarksStore } from "@/stores/bookmarks";
 import { IDBworkflow } from "../idb";
 
-export default {
-  data() {
-    return {
-      store: useStore(),
-    };
-  },
+const bookmarksStore = useBookmarksStore();
+const bookmarks = computed(() => bookmarksStore.idbBookmarks);
 
-  computed: {
-    bookmarks: function () {
-      return this.store.idbBookmarks;
-    },
-  },
+function deletebookmark(id) {
+  IDBworkflow(
+    bookmarksStore.idbBookmarkDbname,
+    bookmarksStore.idbBookmarkVDbver,
+    bookmarksStore.idbBookmarkVDbtable,
+    "readwrite",
+    (storeObj) => {
+      storeObj.delete(id);
 
-  methods: {
-    deletebookmark(id) {
-      IDBworkflow(
-        this.store.idbBookmarkDbname,
-        this.store.idbBookmarkVDbver,
-        this.store.idbBookmarkVDbtable,
-        "readwrite",
-        (store) => {
-          store.delete(id);
+      const bc = new BroadcastChannel(bookmarksStore.idbWatcherBroadcast);
+      bc.postMessage(bookmarksStore.idbBookmarkVDbtable);
+      bc.close();
+    }
+  );
 
-          const bc = new BroadcastChannel(this.store.idbWatcherBroadcast);
-          bc.postMessage(this.store.idbBookmarkVDbtable);
-          bc.close();
-        }
-      );
-        const sensorElement = document.querySelector(`[data-id="${id}"]`);
-        if (sensorElement && sensorElement.classList.contains('sensor-bookmarked')) {
-          sensorElement.classList.remove('sensor-bookmarked');
-        }
-    },
+  const sensorElement = document.querySelector(`[data-id="${id}"]`);
+  if (sensorElement && sensorElement.classList.contains("sensor-bookmarked")) {
+    sensorElement.classList.remove("sensor-bookmarked");
+  }
+}
 
-    showsensor(bookmark) {
-      /* 
-            I'm really sorry for that, but current routing system is very complicated and it seems 
-            very hard to impelement proper router mechanism here, so I just reload - positivecrash */
-      window.location.href = this.getlink(bookmark);
-      location.reload();
-    },
+function showsensor(bookmark) {
+  window.location.href = getlink(bookmark);
+  location.reload();
+}
 
-    getlink(bookmark) {
-      if (bookmark.link && bookmark.geo) {
-        const g = JSON.parse(bookmark.geo);
-        const provider = localStorage.getItem("provider_type") || "remote";
-        return (
-          window.location.origin +
-          "/#/" +
-          provider +
-          "/" +
-          "pm10" +
-          "/" +
-          "20" +
-          "/" +
-          g.lat +
-          "/" +
-          g.lng +
-          "/" +
-          bookmark.link
-        );
-      }
-    },
-  },
-};
+function getlink(bookmark) {
+  if (bookmark.link && bookmark.geo) {
+    const g = JSON.parse(bookmark.geo);
+    const provider = localStorage.getItem("provider_type") || "remote";
+    return (
+      window.location.origin +
+      "/#/" +
+      provider +
+      "/" +
+      "pm10" +
+      "/" +
+      "20" +
+      "/" +
+      g.lat +
+      "/" +
+      g.lng +
+      "/" +
+      bookmark.link
+    );
+  }
+}
+
+onMounted(() => {
+  bookmarksStore.idbBookmarkGet();
+});
 </script>
 
 <style scoped>